+++ diginode.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 #
-#           Name:  DigiNode Dashboard v0.9.1
+#           Name:  DigiNode Dashboard v0.9.2
 #
 #        Purpose:  Monitor and manage the status of you DigiByte Node and DigiAsset Node.
 #          
@@ -57,11 +57,11 @@
 # Local variables will be in lowercase and will exist only within functions
 
 # This variable stores the version number of this release of 'DigiNode Tools'.
-# Wheneve there is a new release, this number gets updated to match the release number on GitHub.
+# Whenever there is a new release, this number gets updated to match the release number on GitHub.
 # The version number should be three numbers seperated by a period
 # Do not change this number or the mechanism for installing updates may no longer work.
-DGNT_VER_LOCAL=0.9.1
-# Last Updated: 2023-10-22
+DGNT_VER_LOCAL=0.9.2
+# Last Updated: 2023-10-23
 
 # This is the command people will enter to run the install script.
 DGNT_SETUP_OFFICIAL_CMD="curl -sSL setup.diginode.tools | bash"
@@ -4169,7 +4169,7 @@ fi
 # FORMAT DIGIFACT BASED ON TERMINAL WIDTH
 
 # regenerate digifact box
-if [ "$generate_digifact_box" = "yes" ]; then
+if [ "$generate_digifact_box" = "yes" ] && [ "$digifact_content" != "" ]; then
     digifact_box_termwidth_title=$(format_bordered_title "$digifact_title" " ║ " " ║ " "terminal_width")
     digifact_box_termwidth_content=$(format_bordered_paragraph "$digifact_content" " ║ " " ║ " "  ║ " "terminal_width")
     generate_digifact_box="no"
@@ -4177,11 +4177,14 @@ fi
 
 # Setup function to display DigiFact dashboard
 display_digifact_termwidth() {
-    echo "$sm_row_digifact_topbar" # "╔" "═" "═" "═" "═" "═" "╗"
-    echo "$digifact_box_termwidth_title"
-    echo "$sm_row_digifact_middle" # "╠" "═" "═" "═" "═" "═" "╣"
-    echo "$digifact_box_termwidth_content"
-    echo "$sm_row_digifact_bottom" # "╚" "═" "═" "═" "═" "═" "╝"
+
+    if [ "$digifact_content" != "" ]; then
+        echo "$sm_row_digifact_topbar" # "╔" "═" "═" "═" "═" "═" "╗"
+        echo "$digifact_box_termwidth_title"
+        echo "$sm_row_digifact_middle" # "╠" "═" "═" "═" "═" "═" "╣"
+        echo "$digifact_box_termwidth_content"
+        echo "$sm_row_digifact_bottom" # "╚" "═" "═" "═" "═" "═" "╝"
+    fi
 } 
 
 # FORMAT LOWCON MESSAGES BASED ON TERMINAL WIDTH
@@ -4559,12 +4562,18 @@ echo "$sm_row_05" # "╚" "═" "╩" "═" "╩" "═" "╝"
 echo "$sm_row_06" # "╔" "═" "═" "═" "╦" "═" "╗"
 
 if [ "$DGNT_UPDATE_AVAILABLE" = "yes" ]; then
+    dgnt_upd_leftcol=" ║ SOFTWARE       ║  "
+    dgnt_upd_maincol="DigiNode Tools $DGNT_VER_LOCAL_DISPLAY"
+
     if [ $term_width -gt 111 ]; then 
-        printf " ║ SOFTWARE       ║  " && printf "%-${col_width_software_dgnt_wide}s %50s %3s\n" "DigiNode Tools $DGNT_VER_LOCAL_DISPLAY" "[ ${txtbgrn}Update Available: DigiNode Tools v$DGNT_VER_RELEASE${txtrst} ]" " ║ "
+        dgnt_upd_rightcol="${dbcol_bgrn}Update Available: DigiNode Tools v$DGNT_VER_RELEASE${dbcol_rst}"
+        db_content_c1_c2_c3f_c4 "$dgnt_upd_leftcol" "$dgnt_upd_maincol" "$dgnt_upd_rightcol"
     elif [ $term_width -gt 97 ]; then
-        printf " ║ SOFTWARE       ║  " && printf "%-${col_width_software}s %35s %3s\n" "DigiNode Tools $DGNT_VER_LOCAL_DISPLAY" "[ ${txtbgrn}Update Available: v$DGNT_VER_RELEASE${txtrst} ]" " ║ "
+        dgnt_upd_rightcol="${dbcol_bgrn}Update Available: v$DGNT_VER_RELEASE${dbcol_rst}"
+        db_content_c1_c2_c3f_c4 "$dgnt_upd_leftcol" "$dgnt_upd_maincol" "$dgnt_upd_rightcol"
     else
-        printf " ║ SOFTWARE       ║  " && printf "%-${col_width_software_narrow}s %27s %3s\n" "DigiNode Tools $DGNT_VER_LOCAL_DISPLAY" "[ ${txtbgrn}Update: v$DGNT_VER_RELEASE${txtrst} ]" " ║ "
+        dgnt_upd_rightcol="${dbcol_bgrn}Update: v$DGNT_VER_RELEASE${dbcol_rst}"
+        db_content_c1_c2_c3f_c4 "$dgnt_upd_leftcol" "$dgnt_upd_maincol" "$dgnt_upd_rightcol"
     fi
 else
     printf " ║ SOFTWARE       ║  " && printf "%-${col_width_software_noupdate}s %-3s\n" "DigiNode Tools $DGNT_VER_LOCAL_DISPLAY" " ║ "
@@ -4572,23 +4581,39 @@ fi
 # printf "  ║               ╠════════════════════════════════════════════════════╣\\n"
 if [ "$DGB_VER_LOCAL" != "" ]; then
     if [ "$DGB_UPDATE_AVAILABLE" = "yes" ]; then
+        sw_upd_leftcol=" ║                ║  "
         if [ "$DGB_PRERELEASE" = "YES" ]; then
+
             if [ $term_width -gt 111 ]; then 
-                printf " ║                ║  " && printf "%-${col_width_software_dgbpr_wide}s %50s %-3s\n" "DigiByte Core v$DGB_VER_LOCAL (Pre-Release)" "[ ${txtbgrn}Update Available: DigiByte Core v$DGB_VER_GITHUB${txtrst} ]" " ║ "
-            elif [ $term_width -gt 97 ]; then
-                printf " ║                ║  " && printf "%-${col_width_software_dgbpr}s %37s %3s\n" "DigiByte Core v$DGB_VER_LOCAL (Pre-Release)" "[ ${txtbgrn}Update Available: v$DGB_VER_GITHUB${txtrst} ]" " ║ "
-            elif [ $term_width -gt 88 ]; then
-                printf " ║                ║  " && printf "%-${col_width_software_dgbpr_lessnarrow}s %35s %3s\n" "DigiByte Core v$DGB_VER_LOCAL (Pre-Release)" "[ ${txtbgrn}Update: v$DGB_VER_GITHUB${txtrst} ]" " ║ "
+                dgb_upd_maincol="DigiByte Core v$DGB_VER_LOCAL (Pre-Release)"
+                dgb_upd_rightcol="${dbcol_bgrn}Update Available: DigiByte Core v$DGB_VER_GITHUB${dbcol_rst}"
+                db_content_c1_c2_c3f_c4 "$sw_upd_leftcol" "$dgb_upd_maincol" "$dgb_upd_rightcol"
+            elif [ $term_width -gt 99 ]; then
+                dgb_upd_maincol="DigiByte Core v$DGB_VER_LOCAL (Pre-Release)"
+                dgb_upd_rightcol="${dbcol_bgrn}Update Available: v$DGB_VER_GITHUB${dbcol_rst}"
+                db_content_c1_c2_c3f_c4 "$sw_upd_leftcol" "$dgb_upd_maincol" "$dgb_upd_rightcol"
+            elif [ $term_width -gt 90 ]; then
+                dgb_upd_maincol="DigiByte Core v$DGB_VER_LOCAL (Pre-Release)"
+                dgb_upd_rightcol="${dbcol_bgrn}Update: v$DGB_VER_GITHUB${dbcol_rst}"
+                db_content_c1_c2_c3f_c4 "$sw_upd_leftcol" "$dgb_upd_maincol" "$dgb_upd_rightcol"
             else
-                printf " ║                ║  " && printf "%-${col_width_software_dgbpr_narrow}s %27s %3s\n" "DigiByte Core v$DGB_VER_LOCAL (PR)" "[ ${txtbgrn}Update: v$DGB_VER_GITHUB${txtrst} ]" " ║ "
+                dgb_upd_maincol="DigiByte Core v$DGB_VER_LOCAL (PR)"
+                dgb_upd_rightcol="${dbcol_bgrn}Update: v$DGB_VER_GITHUB${dbcol_rst}"
+                db_content_c1_c2_c3f_c4 "$sw_upd_leftcol" "$dgb_upd_maincol" "$dgb_upd_rightcol"
             fi    
         else   
             if [ $term_width -gt 111 ]; then 
-                printf " ║                ║  " && printf "%-${col_width_software_dgb_wide}s %50s %-3s\n" "DigiByte Core v$DGB_VER_LOCAL" "[ ${txtbgrn}Update Available: DigiByte Core v$DGB_VER_GITHUB${txtrst} ]" " ║ "
+                dgb_upd_maincol="DigiByte Core v$DGB_VER_LOCAL"
+                dgb_upd_rightcol="${dbcol_bgrn}Update Available: DigiByte Core v$DGB_VER_GITHUB${dbcol_rst}"
+                db_content_c1_c2_c3f_c4 "$sw_upd_leftcol" "$dgb_upd_maincol" "$dgb_upd_rightcol"
             elif [ $term_width -gt 97 ]; then
-                printf " ║                ║  " && printf "%-${col_width_software_dgb}s %37s %3s\n" "DigiByte Core v$DGB_VER_LOCAL" "[ ${txtbgrn}Update Available: v$DGB_VER_GITHUB${txtrst} ]" " ║ "
+                dgb_upd_maincol="DigiByte Core v$DGB_VER_LOCAL"
+                dgb_upd_rightcol="${dbcol_bgrn}Update Available: v$DGB_VER_GITHUB${dbcol_rst}"
+                db_content_c1_c2_c3f_c4 "$sw_upd_leftcol" "$dgb_upd_maincol" "$dgb_upd_rightcol"
             else
-                printf " ║                ║  " && printf "%-${col_width_software_dgb_narrow}s %27s %3s\n" "DigiByte Core v$DGB_VER_LOCAL" "[ ${txtbgrn}Update: v$DGB_VER_GITHUB${txtrst} ]" " ║ "
+                dgb_upd_maincol="DigiByte Core v$DGB_VER_LOCAL"
+                dgb_upd_rightcol="${dbcol_bgrn}Update: v$DGB_VER_GITHUB${dbcol_rst}"
+                db_content_c1_c2_c3f_c4 "$sw_upd_leftcol" "$dgb_upd_maincol" "$dgb_upd_rightcol"
             fi   
         fi
     else