+++ docs/release_notes.md
@@ -1,5 +1,10 @@
 ## DigiNode Tools Release Notes
 
+DigiNode Tools v0.9.2 - 2023-10-23
+- Fix: Dashboard spacing when there is an update available in the Software section
+- Fix: Don't check the disk usage of IPFS if it is not present.
+- Fix: Permissions issue with digifacts.json file
+
 DigiNode Tools v0.9.1 - 2023-10-23
 - Fix: Error detecting testnet disk usage when no testnet4 folder exists.
 - Fix: Error with DigiByte update checker when version numbers are the same. It should hopefully be able to correctly identify updates now.